backend/main.py
"""
AI Education Agent - Main FastAPI Application
Provides REST API endpoints for the personalized learning system
"""

from fastapi import FastAPI, HTTPException, Depends, status
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from pydantic import BaseModel
from typing import List, Optional, Dict, Any
import uvicorn
from datetime import datetime
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Initialize FastAPI app
app = FastAPI(
    title="AI Education Agent",
    description="Personalized learning system for government schools",
    version="1.0.0",
    docs_url="/docs",
    redoc_url="/redoc"
)

# CORS middleware for frontend integration
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000", "http://localhost:3001"],  # React dev server
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Security
security = HTTPBearer()

# Pydantic models for API requests/responses
class StudentProfile(BaseModel):
    student_id: str
    name: str
    grade: int
    school_id: str
    language_preference: str = "english"
    learning_style: Optional[str] = None

class AcademicRecordInput(BaseModel):
    student_id: str
    subject: str
    topic: Optional[str] = None
    score: float
    max_score: float
    assessment_type: str
    assessment_date: datetime
    academic_year: str

class LearningSessionRequest(BaseModel):
    student_id: str
    subject: str
    topic: str
    session_type: str = "lesson"

class DoubtQueryRequest(BaseModel):
    student_id: str
    question: str
    context: Optional[str] = None
    subject: Optional[str] = None
    topic: Optional[str] = None

class WorksheetRequest(BaseModel):
    student_id: str
    subject: str
    topic: str
    difficulty_level: int = 1
    num_problems: int = 5

# API Routes

@app.get("/")
async def root():
    """Health check endpoint"""
    return {
        "message": "AI Education Agent API",
        "status": "active",
        "version": "1.0.0",
        "timestamp": datetime.utcnow().isoformat()
    }

@app.get("/health")
async def health_check():
    """Detailed health check"""
    return {
        "status": "healthy",
        "database": "connected",  # TODO: Add actual DB health check
        "ai_services": "available",  # TODO: Add AI service health check
        "timestamp": datetime.utcnow().isoformat()
    }

# Student Management Endpoints

@app.post("/api/students", response_model=Dict[str, Any])
async def create_student(student: StudentProfile):
    """Create a new student profile"""
    # TODO: Implement database integration
    return {
        "message": "Student profile created successfully",
        "student_id": student.student_id,
        "status": "created"
    }

@app.get("/api/students/{student_id}")
async def get_student(student_id: str):
    """Get student profile and learning analytics"""
    # TODO: Implement database query
    return {
        "student_id": student_id,
        "name": "Sample Student",
        "grade": 8,
        "current_level": 75.5,
        "strengths": ["Mathematics", "Science"],
        "weaknesses": ["English Grammar", "History"],
        "learning_style": "visual"
    }

@app.post("/api/students/{student_id}/academic-records")
async def add_academic_record(student_id: str, record: AcademicRecordInput):
    """Add academic performance record"""
    # TODO: Implement database storage and analysis
    return {
        "message": "Academic record added successfully",
        "student_id": student_id,
        "analysis": "Performance analysis updated"
    }

# Learning Session Endpoints

@app.post("/api/learning-sessions")
async def start_learning_session(session_request: LearningSessionRequest):
    """Start a personalized learning session"""
    # TODO: Implement AI-powered content personalization
    return {
        "session_id": "session_123",
        "student_id": session_request.student_id,
        "personalized_content": {
            "difficulty_level": 3,
            "learning_style_adaptation": "visual",
            "content_blocks": [
                {
                    "type": "explanation",
                    "title": f"Introduction to {session_request.topic}",
                    "content": "Personalized explanation based on student's level..."
                },
                {
                    "type": "example",
                    "title": "Practical Example",
                    "content": "Real-world example adapted to student's interests..."
                }
            ]
        }
    }

@app.get("/api/learning-sessions/{session_id}")
async def get_learning_session(session_id: str):
    """Get learning session details and progress"""
    # TODO: Implement session tracking
    return {
        "session_id": session_id,
        "progress": 65.0,
        "time_spent_minutes": 25,
        "concepts_covered": ["Basic concepts", "Applications"],
        "next_recommended_action": "practice_worksheet"
    }

# Worksheet Generation Endpoints

@app.post("/api/worksheets/generate")
async def generate_worksheet(worksheet_request: WorksheetRequest):
    """Generate personalized worksheet based on student's level"""
    # TODO: Implement AI-powered worksheet generation
    return {
        "worksheet_id": "worksheet_456",
        "title": f"{worksheet_request.subject} - {worksheet_request.topic} Practice",
        "difficulty_level": worksheet_request.difficulty_level,
        "estimated_time_minutes": 30,
        "problems": [
            {
                "id": 1,
                "type": "multiple_choice",
                "question": "Sample question adapted to student's level...",
                "options": ["Option A", "Option B", "Option C", "Option D"],
                "difficulty": worksheet_request.difficulty_level
            },
            {
                "id": 2,
                "type": "short_answer",
                "question": "Explain the concept with an example...",
                "difficulty": worksheet_request.difficulty_level
            }
        ],
        "learning_objectives": [
            f"Understand basic concepts of {worksheet_request.topic}",
            f"Apply {worksheet_request.topic} in practical scenarios"
        ]
    }

@app.post("/api/worksheets/{worksheet_id}/submit")
async def submit_worksheet(worksheet_id: str, answers: Dict[str, Any]):
    """Submit worksheet answers and get AI feedback"""
    # TODO: Implement answer evaluation and feedback generation
    return {
        "worksheet_id": worksheet_id,
        "score": 85.0,
        "total_questions": len(answers.get("answers", [])),
        "correct_answers": 4,
        "feedback": {
            "overall": "Great job! You've shown good understanding of the concepts.",
            "detailed": [
                {"question_id": 1, "correct": True, "feedback": "Excellent reasoning!"},
                {"question_id": 2, "correct": False, "feedback": "Consider reviewing the concept of..."}
            ]
        },
        "next_recommendations": [
            "Practice more problems on similar topics",
            "Review the concept that was challenging"
        ]
    }

# Doubt Clearing Endpoints

@app.post("/api/doubts/ask")
async def ask_doubt(doubt_query: DoubtQueryRequest):
    """Submit a doubt/question for AI-powered resolution"""
    # TODO: Implement AI-powered doubt clearing
    return {
        "query_id": "doubt_789",
        "student_id": doubt_query.student_id,
        "question": doubt_query.question,
        "ai_response": {
            "type": "explanation",
            "content": "Here's a detailed explanation of your question...",
            "examples": ["Example 1: ...", "Example 2: ..."],
            "follow_up_questions": [
                "Would you like me to explain this with a different approach?",
                "Do you want to see more examples?"
            ]
        },
        "confidence_score": 0.92,
        "related_topics": ["Topic A", "Topic B"]
    }

@app.get("/api/doubts/{query_id}")
async def get_doubt_resolution(query_id: str):
    """Get doubt resolution details"""
    # TODO: Implement doubt tracking
    return {
        "query_id": query_id,
        "status": "resolved",
        "resolution_time_minutes": 2,
        "student_satisfaction": 4,
        "follow_up_needed": False
    }

# Analytics and Reporting Endpoints

@app.get("/api/analytics/student/{student_id}")
async def get_student_analytics(student_id: str, period: str = "monthly"):
    """Get comprehensive learning analytics for a student"""
    # TODO: Implement analytics calculation
    return {
        "student_id": student_id,
        "period": period,
        "metrics": {
            "total_learning_time_hours": 45.5,
            "sessions_completed": 28,
            "average_score": 78.5,
            "improvement_rate": 12.3,
            "concepts_mastered": 15,
            "areas_for_improvement": ["Grammar", "Problem Solving"]
        },
        "progress_chart": {
            "dates": ["2024-01-01", "2024-01-08", "2024-01-15"],
            "scores": [65, 72, 78]
        },
        "recommendations": [
            "Focus more on grammar exercises",
            "Increase practice time for problem-solving"
        ]
    }

@app.get("/api/analytics/class/{grade}/{school_id}")
async def get_class_analytics(grade: int, school_id: str):
    """Get class-level analytics for teachers"""
    # TODO: Implement class analytics
    return {
        "grade": grade,
        "school_id": school_id,
        "class_metrics": {
            "total_students": 35,
            "active_students": 32,
            "average_class_score": 74.2,
            "completion_rate": 89.5
        },
        "subject_performance": {
            "Mathematics": 78.5,
            "Science": 76.2,
            "English": 71.8,
            "Social Studies": 73.1
        },
        "at_risk_students": [
            {"student_id": "STU001", "risk_level": "high", "reason": "Low engagement"},
            {"student_id": "STU015", "risk_level": "medium", "reason": "Declining scores"}
        ]
    }

# Run the application
if __name__ == "__main__":
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="info"
    )